adds endpoints for data download/upload

+++ src/ploomber/cloud/api.py
@@ -1,7 +1,7 @@
+import cgi
 import sys
 from pathlib import Path
-from urllib.request import urlretrieve
-from urllib import parse
+from urllib.request import urlretrieve, urlopen
 from concurrent.futures import ThreadPoolExecutor, as_completed
 import os
 from glob import glob
@@ -15,18 +15,26 @@ import requests
 import humanize
 
 from ploomber.table import Table
+from ploomber.cloud import io
 
 HOST = os.environ.get(
     'PLOOMBER_CLOUD_HOST',
     'https://8lfxtyhlx2.execute-api.us-east-1.amazonaws.com/api/')
 
 
-def _download_file(url):
-    # remove leading /
-    path = Path(parse.urlparse(url).path[1:])
-    path.parent.mkdir(exist_ok=True, parents=True)
-    print(f'Downloading {path}')
-    urlretrieve(url, path)
+def _download_file(url, skip_if_exists=False):
+    remotefile = urlopen(url)
+    _, params = cgi.parse_header(remotefile.info()['Content-Disposition'])
+    path = params["filename"]
+
+    if skip_if_exists and Path(path).exists():
+        print(f'{path} exists, skipping...')
+    else:
+        Path(path).parent.mkdir(exist_ok=True, parents=True)
+        print(f'Downloading {path}')
+        urlretrieve(url, path)
+
+    return path
 
 
 def _request_factory(method):
@@ -259,3 +267,35 @@ def upload_project(force, github_number, github_owner, github_repo, verbose):
     # TODO: if anything fails after runs_new, update the status to error
 
     return runid
+
+
+@auth_header
+def upload_data(headers, path):
+    key = Path(path).name
+
+    create = _post(f"{HOST}/upload/data/create",
+                   headers=headers,
+                   json=dict(key=key, n_parts=io.n_parts(path))).json()
+
+    gen = io.UploadJobGenerator(path,
+                                key=key,
+                                upload_id=create['upload_id'],
+                                links=create['urls'])
+
+    click.echo('Uploading...')
+    parts = gen.upload()
+
+    _post(f"{HOST}/upload/data/complete",
+          headers=headers,
+          json=dict(key=key, parts=parts,
+                    upload_id=create['upload_id'])).json()
+
+    # print snippet showing how to download it in the pipeline
+
+
+@auth_header
+def download_data(headers, key):
+    response = _post(f"{HOST}/download/data",
+                     headers=headers,
+                     json=dict(key=key))
+    return response
